backend/requirements.txt
aerich==0.7.1
aiomysql==0.1.1
aiosmtplib==1.1.7
aiosqlite==0.17.0
amqp==5.1.1
anyio==3.6.2
async-timeout==4.0.2
bcrypt==4.0.1
billiard==3.6.4.0
blinker==1.5
boto3==1.26.32
botocore==1.29.32
celery==5.2.7
certifi==2022.12.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
dictdiffer==0.9.0
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
fastapi==0.88.0
fastapi-mail==1.2.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.2
httpx==0.23.1
idna==3.4
iso8601==1.1.0
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.2.4
MarkupSafe==2.1.1
passlib==1.7.4
Pillow==9.3.0
prompt-toolkit==3.0.36
pyasn1==0.4.8
pydantic==1.10.2
PyMySQL==1.0.2
pypika-tortoise==0.1.6
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.7
redis==4.4.0
reportlab==3.6.12
rfc3986==1.5.0
rsa==4.9
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
starlette==0.21.0
tomlkit==0.11.6
tortoise-orm==0.19.2
typing-extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
vine==5.0.0
wcwidth==0.2.5
